+++ config/constants.php
@@ -4,8 +4,10 @@ return [
     'DEUDAS_TECNICA_MAIL_GROUP' => 2,
     'CAMBIO_ESTADO_PROVEEDOR_MAIL_GROUP' => 3,
     'USO_HORARIO_ARG' => 'America/Argentina/Buenos_Aires',
-    'SLAM_VERSION' => '0.7.10',
+    'SLAM_VERSION' => '0.7.11',
     'CLIENT_USER' => 'ubnt',
     'CLIENT_PASS' => 'S1mon3d@',
-    'LAN_SEGMENT' => '10.10.0.0/16'
+    'LAN_SEGMENT' => '10.10.0.0/16',
+    'IP_INICIAL' => '10.10.20.1',
+    'IP_ACTUAL' => ''
 ];