+++ database/migrations/2022_02_02_231658_create_variables_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateVariablesTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('variables', function (Blueprint $table) {
+            $table->id();
+            $table->timestamps();
+            $table->ipAddress('ip_inical');
+            $table->ipAddress('ip_actual');
+            $table->ipAddress('ip_final');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('variables');
+    }
+}